refactor(openapi-validator): validate via @scalar/json-schema-validator
openapi-validator now delegates schema validation and error prettifying to the shared @scalar/json-schema-validator engine, keeping only the OpenAPI-specific parts: version detection, the schema registry, and the path-parameter semantic checks. Drops the direct ajv/ajv-draft-04/ajv-formats dependencies. Public behaviour is unchanged.
