+++ packages/openapi-validator/package.json
@@ -44,10 +44,8 @@
   "sideEffects": false,
   "dependencies": {
     "@scalar/helpers": "workspace:*",
+    "@scalar/json-schema-validator": "workspace:*",
     "@scalar/types": "workspace:*",
-    "ajv": "^8.17.1",
-    "ajv-draft-04": "^1.0.0",
-    "ajv-formats": "^3.0.1",
     "yaml": "catalog:*"
   },
   "devDependencies": {
